Move get_ipoib_addr() to vys_get_ipoib_addr()
